+++ aws_swiffer/resources/s3/Bucket.py
@@ -1,9 +1,13 @@
 import os
+from typing import TYPE_CHECKING
 
 import botocore.exceptions
 
 from aws_swiffer.resources.IResource import IResource
-from aws_swiffer.utils import get_resource, get_logger, ask_delete_confirm
+from aws_swiffer.utils import get_resource, get_logger
+
+if TYPE_CHECKING:
+    from aws_swiffer.utils.context import ExecutionContext
 
 logger = get_logger(os.path.basename(__file__))
 
@@ -17,40 +21,51 @@ class Bucket(IResource):
         self.s3 = get_resource('s3', self.region)
         self.bucket = self.s3.Bucket(self.name)
 
-    def remove(self, clear_only: bool = False):
-        logger.info(f"Trying to delete resource: {self.arn}")
+    def remove(self, context: 'ExecutionContext' = None):
+        prefix = context.log_prefix() if context else ""
+        logger.info(f"{prefix}Trying to delete resource: {self.arn}")
 
-        delete = ask_delete_confirm(self.name)
-        if delete:
-            try:
-                self.clean()
-                try:
-                    bucket_website = self.s3.BucketWebsite(self.name)
-                    logger.info('Trying to delete website configuration')
-                    bucket_website.delete()
-                    logger.info('Website configuration deleted')
-                except botocore.exceptions.ClientError as e:
-                    logger.debug(e)
-                response = self.bucket.delete()
-                logger.debug(response)
-                logger.info(f"Resource deleted: {self.arn}")
-            except botocore.exceptions.ClientError as e:
-                if e.response.get('Error', {}).get('Code') == 'NoSuchBucket':
-                    logger.info("Bucket not found")
-                logger.error(f"Cannot delete resource: {self.arn}")
-                logger.debug(e)
-        else:
+        if not self._should_proceed(context, "delete bucket"):
             logger.info("Delete skipped")
+            return
+        
+        if context and context.dry_run:
+            logger.info(f"{prefix}Would delete bucket: {self.name}")
+            return
 
-    def clean(self):
         try:
-            logger.info(f'Start clean for: {self.arn}')
-            logger.info('Trying to delete old versions')
+            self.clean(context)
+            try:
+                bucket_website = self.s3.BucketWebsite(self.name)
+                logger.info(f'{prefix}Trying to delete website configuration')
+                bucket_website.delete()
+                logger.info(f'{prefix}Website configuration deleted')
+            except botocore.exceptions.ClientError as e:
+                logger.debug(e)
+            response = self.bucket.delete()
+            logger.debug(response)
+            logger.info(f"{prefix}Resource deleted: {self.arn}")
+        except botocore.exceptions.ClientError as e:
+            if e.response.get('Error', {}).get('Code') == 'NoSuchBucket':
+                logger.info("Bucket not found")
+            logger.error(f"Cannot delete resource: {self.arn}")
+            logger.debug(e)
+
+    def clean(self, context: 'ExecutionContext' = None):
+        prefix = context.log_prefix() if context else ""
+        
+        if context and context.dry_run:
+            logger.info(f"{prefix}Would clean bucket: {self.name}")
+            return
+        
+        try:
+            logger.info(f'{prefix}Start clean for: {self.arn}')
+            logger.info(f'{prefix}Trying to delete old versions')
             self.bucket.object_versions.delete()
-            logger.info('Old file versions deleted ')
+            logger.info(f'{prefix}Old file versions deleted ')
         except botocore.exceptions.ClientError as e:
             logger.debug(e)
         else:
-            logger.info(f"Start delete of all objects in bucket")
+            logger.info(f"{prefix}Start delete of all objects in bucket")
             self.bucket.objects.all().delete()
-            logger.info(f"Delete of all objects completed")
+            logger.info(f"{prefix}Delete of all objects completed")
